+++ CHANGELOG.md
@@ -118,6 +118,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (GL #1105) Upload modal restricted to uploading to current container, or creating a new one in main container view
 - (GL #1028) Disallow uploading files with size 0
 - Use common base class for database connections
+- (GL #1153) Display decimals with single digit file sizes
 
 ### Fixed